Added Further question component and streaming fix
